+++ sdk/core/iOS/MoEReactBridge/MoEngageReactConstants.h
@@ -47,3 +47,4 @@ extern NSString* const kPushTokenGenerated;
 extern NSString* const kPushClicked;
 extern NSString* const kPermissionResult;
 extern NSString* const kLogoutComplete;
+extern NSString* const kAuthenticationError;