+++ sdk/core/src/models/MoEAuthenticationType.ts
@@ -1,0 +1,8 @@
+/**
+ * Type of authentication used to authenticate the SDK's network requests.
+ *
+ * @since 12.10.0
+ */
+export enum MoEAuthenticationType {
+    JWT = "JWT"
+}