+++ sdk/core/src/utils/MoEEventHandlerHelper.ts
@@ -7,8 +7,8 @@ import {
 } from "../moeParser/MoEInAppParser";
 import {getMoEPushPayload, getMoEPushToken} from "../moeParser/MoEPushNotificationParser";
 import {MOE_DATA, ACCOUNT_META, MOE_PAYLOAD} from "./MoEConstants";
-import {PERMISSION_RESULT, LOGOUT_COMPLETE} from "..";
-import {getPermissionResult, getLogoutCompleteData} from "../moeParser/MoEngagePayloadParser";
+import {PERMISSION_RESULT, LOGOUT_COMPLETE, AUTHENTICATION_ERROR} from "..";
+import {getPermissionResult, getLogoutCompleteData, getAuthenticationErrorData} from "../moeParser/MoEngagePayloadParser";
 
 export function executeHandler(
     handler: Function,
@@ -28,6 +28,9 @@ export function executeHandler(
             } else if (type == LOGOUT_COMPLETE) {
                 let logoutData = getLogoutCompleteData(notificationPayload);
                 logoutData != undefined ? handler(logoutData) : null;
+            } else if (type == AUTHENTICATION_ERROR) {
+                let authError = getAuthenticationErrorData(notificationPayload);
+                authError != undefined ? handler(authError) : null;
             } else {
                 const accountMeta = notificationPayload[ACCOUNT_META];
                 if (accountMeta != undefined && isValidObject(accountMeta)) {